✨ restore and refactor Singlefull view
