+++ src/views/Single/Single.js
@@ -1,23 +1,8 @@
 import React, { Component } from 'react';
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
-import PropTypes from 'prop-types';
-import ReactMarkdown from 'react-markdown';
-import Item from '../../components/Item/Item';
-
-import { getComponentMarkup, getVariantMarkup } from '../../actions/atomic';
 
 import './Single.css';
 
 class Single extends Component {
-  constructor() {
-    super();
-
-    this.state = {
-      component: {},
-    }
-  }
-
   componentWillMount() {
     this.getContent(this.props);
   }
@@ -40,67 +25,6 @@ class Single extends Component {
 
     this.setState({ component });
   }
-
-  render() {
-    const variants = this.state.component.variants.length > 0 && (
-      <div>
-        {this.state.component.variants.map((variant, key) => {
-          return (
-            <Item
-              wrapper={this.state.component.wrapper || ''}
-              background={this.state.component.background}
-              key={key}
-              title={variant.title}
-              slug={`tlbx-${this.state.component.slug}-${variant.slug}`}
-              fullUrl={`/${this.props.location.pathname.split('/')[1]}/${this.props.match.params.slug}/${variant.slug}/full`}
-            >
-              {variant.content || ''}
-            </Item>
-          );
-        })}
-      </div>
-    );
-
-    return (
-      <div>
-        <h1 className="tlbx-h1">{this.state.component.title}</h1>
-        {/* {this.state.component.notes && (
-          <div className="tlbx-notes">
-            <ReactMarkdown source={this.state.component.notes} />
-          </div>
-        )} */}
-
-        <Item
-          wrapper={this.state.component.wrapper || ''}
-          background={this.state.component.background}
-          slug={`tlbx-${this.state.component.slug}`}
-          fullUrl={`/${this.props.location.pathname.split('/')[1]}/${this.props.match.params.slug}/full`}
-        >
-          {this.state.component.content || ''}
-        </Item>
-
-        {variants}
-      </div>
-    );
-  }
 }
 
-Single.propTypes = {
-  components: PropTypes.object,
-};
-
-function mapState(state) {
-  return {
-    atomic: state.atomic,
-    navigation: state.navigation,
-  };
-}
-
-function mapDispatch(dispatch) {
-  return bindActionCreators({
-    getComponentMarkup,
-    getVariantMarkup,
-  }, dispatch);
-}
-
-export default connect(mapState, mapDispatch)(Single);
+export default Single;
