+++ src/views/Single/Single.scss
@@ -4,3 +4,21 @@
   max-width: 700px;
   line-height: 1.55em;
 }
+
+.tlbx-single-full {
+  min-height: 100vh;
+
+  .tlbx-item-preview {
+    border: 0;
+  }
+
+  h1 {
+    font-size: 1rem;
+    margin: 0;
+  }
+}
+
+.tlbx-single-full-intro {
+  padding: 1rem;
+  background-color: $gray-lighter;
+}
