--- src/views/SingleFull/SingleFull.scss
@@ -1,19 +1,0 @@
-@import '../../variables.scss';
-
-.tlbx-single-full {
-  min-height: 100vh;
-
-  .tlbx-item-preview {
-    border: 0;
-  }
-
-  h1 {
-    font-size: 1rem;
-    margin: 0;
-  }
-}
-
-.tlbx-single-full-intro {
-  padding: 1rem;
-  background-color: $gray-lighter;
-}
